requirements.txt
numpy>=1.19.2
pandas>=1.2.1
scikit-learn>=0.24.1
seqeval>=1.2.2
transformers>=4.2.2
datasets>=1.2.1
pytorch-lightning==1.1.8
python-box>=5.3.0
pyyaml>=5.3
dataclasses>=0.7